+++ beltradar/api/survey.py
@@ -10,7 +10,7 @@ from allianceauth.services.hooks import get_extension_logger
 # AA Belt Radar
 from beltradar import __title__
 from beltradar.api import schema
-from beltradar.api.helpers.core import get_owner_or_none
+from beltradar.api.helpers.core import get_owner_or_none, get_public_id_or_none
 from beltradar.api.helpers.icons import (
     get_snapshot_delete_button,
     get_survey_manage_action_icons,
@@ -27,11 +27,11 @@ class BeltRadarSurveyApiEndpoints:
 
     def __init__(self, api: NinjaAPI):
         @api.get(
-            "survey/{public_id}/view/",
+            "view/session/{public_id}/",
             response={200: schema.BeltSurveySessionSchema, 403: str},
             tags=self.tags,
         )
-        def get_survey_sessions(request, public_id: str):
+        def get_survey_session(request, public_id: str):
             """Get survey session details for a specific public_id."""
             if not request.user.has_perm("beltradar.basic_access"):
                 return 403, _("You do not have permission to access this resource.")
@@ -52,16 +52,13 @@ class BeltRadarSurveyApiEndpoints:
             return 200, survey_session_data
 
         @api.get(
-            "survey/view/my-sessions/{character_id}/",
+            "view/my-sessions/{character_id}/",
             response={200: list[schema.BeltSurveySessionSchema], 403: dict, 404: dict},
             tags=self.tags,
         )
         def get_my_sessions(request, character_id: int):
             """Get all survey sessions for the current user."""
-            perms, owner = get_owner_or_none(request=request, character_id=character_id)
-
-            if owner is None:
-                return 404, {"error": _("User not found.")}
+            perms = get_owner_or_none(request=request, character_id=character_id)[0]
 
             if perms is False:
                 return 403, {"error": _("Permission Denied.")}
@@ -86,7 +83,7 @@ class BeltRadarSurveyApiEndpoints:
             return 200, survey_list
 
         @api.get(
-            "survey/{public_id}/view/entries/",
+            "view/session/{public_id}/entries/",
             response={200: list[schema.OreSchemaList], 403: str},
             tags=self.tags,
         )
@@ -138,8 +135,61 @@ class BeltRadarSurveyApiEndpoints:
                 )
             return 200, survey_list
 
+        @api.get(
+            "view/session/{public_id}/snapshot/last_entry/",
+            response={200: schema.OreSchemaList, 403: str},
+            tags=self.tags,
+        )
+        def get_survey_entry(request, public_id: str):
+            """Get all survey entries for the current user."""
+            if not request.user.has_perm("beltradar.basic_access"):
+                return 403, _("You do not have permission to access this resource.")
+
+            # Check if the survey session exists
+            session = BeltSurveySession.objects.filter(public_id=public_id).first()
+            if not session:
+                return 403, _("Survey session not found or not public.")
+
+            # Get the most recent survey entry for this session (if any)
+            entries = (
+                session.br_entries.select_related("eve_type")
+                .order_by("-timestamp")
+                .filter(snapshot=session.last_entry_snapshot())
+            )
+
+            # Create a list of survey entries for the current user, ordered by timestamp
+            snapshot_list: list[schema.OreSchema] = []
+            for entry in entries:
+                ore_data = schema.OreSchema(
+                    portrait=(
+                        get_icon_render_url(type_id=entry.eve_type.id, as_html=True)
+                        if entry.eve_type
+                        else None
+                    ),
+                    name=entry.eve_type.name if entry.eve_type else "Unknown",
+                    units=entry.units or 0,
+                    volume_m3=entry.volume_left or 0,
+                    price_isk=entry.price or 0.0,
+                    price_compressed=entry.price_compressed or None,
+                    timestamp=entry.timestamp,
+                    snapshot=entry.snapshot,
+                )
+                snapshot_list.append(ore_data)
+            return 200, schema.OreSchemaList(
+                snapshot=entries[0].snapshot,
+                timestamp=entries[0].timestamp,
+                entries=snapshot_list,
+                delete_html=str(
+                    get_snapshot_delete_button(
+                        request=request,
+                        public_id=entries[0].session.public_id,
+                        snapshot=entries[0].snapshot,
+                    )
+                ),
+            )
+
         @api.post(
-            "survey/{public_id}/manage/delete/",
+            "session/{public_id}/manage/delete/",
             response={200: dict, 403: dict, 404: dict},
             tags=self.tags,
         )
@@ -164,9 +214,9 @@ class BeltRadarSurveyApiEndpoints:
                 return 404, {"error": msg}
 
             # Check if the user has permission to delete this survey session
-            perms = get_owner_or_none(
+            perms = get_public_id_or_none(
                 request=request,
-                character_id=session.owner.profile.main_character.character_id,
+                public_id=public_id,
             )[0]
             if not perms:
                 msg = _("Permission Denied.")
@@ -179,7 +229,7 @@ class BeltRadarSurveyApiEndpoints:
             return 200, {"success": True, "message": msg}
 
         @api.post(
-            "survey/{public_id}/snapshot/{snapshot}/manage/delete/",
+            "session/{public_id}/snapshot/{snapshot}/manage/delete/",
             response={200: dict, 403: dict, 404: dict},
             tags=self.tags,
         )